Change float to double (f64)

+++ src/lib.rs
@@ -6,6 +6,6 @@ pub mod prelude;
 
 
 pub trait NNPS {
-    fn register_particles_to_nnps(&mut self, x: &[f32], y: &[f32], z: &[f32]);
-    fn get_neighbours(&self, x: f32, y: f32, z: f32) -> Vec<usize>;
+    fn register_particles_to_nnps(&mut self, x: &[f64], y: &[f64], z: &[f64]);
+    fn get_neighbours(&self, x: f64, y: f64, z: f64) -> Vec<usize>;
 }
